fix: eslint errors repared

+++ src/app/api/vehiculos/[uuid]/route.ts
@@ -1,40 +1,50 @@
 import { NextResponse } from 'next/server'
 import prisma from '../../../../lib/prisma'
 
+async function findOne(uuid: string) {
+  const vehiculo = await prisma.vehiculo.findUnique({
+    where: { uuid },
+    include: {
+      areaTrabajo: {
+        select: {
+          uuid: true,
+          nombre: true,
+          descripcion: true,
+        },
+      },
+      chofer: {
+        select: {
+          nombre: true,
+          ci: true,
+          licencia: true,
+        },
+      },
+      tarjeta: {
+        select: {
+          uuid: true,
+          numero: true,
+          estado: true,
+          fecha_vencimiento: true,
+        },
+      },
+    },
+  })
+
+  return vehiculo;
+}
+
+type tParams = Promise<{ uuid: string }>;
+
 // GET vehicle by uuid
 export async function GET(
   request: Request,
-  { params }: { params: { uuid: string } }
+  { params }: { params: tParams }
 ) {
   try {
-    const vehiculo = await prisma.vehiculo.findUnique({
-      where: { uuid: params.uuid },
-      include: {
-        areaTrabajo: {
-          select: {
-            uuid: true,
-            nombre: true,
-            descripcion: true,
-          },
-        },
-        chofer: {
-          select: {
-            uuid: true,
-            nombre: true,
-            ci: true,
-            licencia: true,
-          },
-        },
-        tarjeta: {
-          select: {
-            uuid: true,
-            numero: true,
-            estado: true,
-            fecha_vencimiento: true,
-          },
-        },
-      },
-    })
+
+    const { uuid } = await params;
+
+    const vehiculo = await findOne(uuid); 
 
     if (!vehiculo) {
       return NextResponse.json({ error: 'Vehicle not found' }, { status: 404 })
@@ -42,20 +52,29 @@ export async function GET(
 
     return NextResponse.json(vehiculo)
   } catch (error) {
-    return NextResponse.json({ error: 'Error fetching vehicle' }, { status: 500 })
+    return NextResponse.json({ error: 'Error fetching vehicle', message: error }, { status: 500 })
   }
 }
 
 // PUT update vehicle
-export async function PUT(
+export async function PATCH(
   request: Request,
-  { params }: { params: { uuid: string } }
+  { params }: { params: tParams }
 ) {
   try {
-    const { chapa, marca, tipo, jefe, isAvailable, areaTrabajoUuid, choferUuid } = await request.json()
+
+    const { uuid } = await params;
+
+    const vehiculoExists = await findOne(uuid); 
+
+    if (!vehiculoExists) {
+      return NextResponse.json({ error: 'Vehicle not found' }, { status: 404 })
+    }
+
+    const { chapa, marca, tipo, jefe, isAvailable, areaTrabajoUuid, choferCI } = await request.json()
     
     const vehiculo = await prisma.vehiculo.update({
-      where: { uuid: params.uuid },
+      where: { uuid },
       data: {
         chapa,
         marca,
@@ -63,7 +82,7 @@ export async function PUT(
         jefe,
         isAvailable,
         areaTrabajoUuid,
-        choferUuid,
+        choferCI,
       },
       include: {
         areaTrabajo: {
@@ -75,7 +94,6 @@ export async function PUT(
         },
         chofer: {
           select: {
-            uuid: true,
             nombre: true,
             ci: true,
             licencia: true,
@@ -94,22 +112,31 @@ export async function PUT(
     
     return NextResponse.json(vehiculo)
   } catch (error) {
-    return NextResponse.json({ error: 'Error updating vehicle' }, { status: 500 })
+    return NextResponse.json({ error: 'Error updating vehicle', message: error }, { status: 500 })
   }
 }
 
 // DELETE vehicle
 export async function DELETE(
   request: Request,
-  { params }: { params: { uuid: string } }
+  { params }: { params: tParams }
 ) {
   try {
+
+    const { uuid } = await params;
+
+    const vehiculo = await findOne(uuid); 
+
+    if (!vehiculo) {
+      return NextResponse.json({ error: 'Vehicle not found' }, { status: 404 })
+    }
+
     await prisma.vehiculo.delete({
-      where: { uuid: params.uuid },
+      where: { uuid },
     })
     
     return NextResponse.json({ message: 'Vehicle deleted successfully' })
   } catch (error) {
-    return NextResponse.json({ error: 'Error deleting vehicle' }, { status: 500 })
+    return NextResponse.json({ error: 'Error deleting vehicle', message: error }, { status: 500 })
   }
 }
